extract merge layout from task layout
this split will ease the introduction of a deleteLayout

while reviewing router update, keep in mind that the router have 2 ways to access a taskLayout : by taskId and by a pluralized entitiesType (ie. /tasks/humans)"
